+++ ModelViewer/MeshGenerator.h
@@ -5,7 +5,7 @@
 class MeshGenerator {
 public:
     static MeshData GenerateCube();
-    static MeshData GenerateAxisGrid(const int length, const int unit, Vector3 gridColor,
+    static MeshData GenerateAxisGrid(const float length, const int unit, Vector3 gridColor,
                                          float axisBrightness);
     static MeshData Square();
     static MeshData GenerateXZSquare(const float length);
