+++ ModelViewer/Renderer.cpp
@@ -54,10 +54,10 @@ void Renderer::DrawIndexed(Model &model) {
 
     std::vector<ID3D11Buffer *> buffers = {cameraTransformBuffer.Get(),
                                            model.transformationBuffer.Get()};
-    context->VSSetConstantBuffers(0, buffers.size(), buffers.data());
-    context->GSSetConstantBuffers(0, buffers.size(), buffers.data());
-    context->HSSetConstantBuffers(0, buffers.size(), buffers.data());
-    context->DSSetConstantBuffers(0, buffers.size(), buffers.data());
+    context->VSSetConstantBuffers(0, UINT(buffers.size()), buffers.data());
+    context->GSSetConstantBuffers(0, UINT(buffers.size()), buffers.data());
+    context->HSSetConstantBuffers(0, UINT(buffers.size()), buffers.data());
+    context->DSSetConstantBuffers(0, UINT(buffers.size()), buffers.data());
 
     context->OMSetRenderTargets(1, rawRenderTargetView.GetAddressOf(),
                                 depthStencilView.Get());
